+++ src/features/mint-nft-drop/Steps/Finished/Finished.module.scss
