+++ src/features/mint-nft-drop/Steps/Info/Info.tsx
