+++ src/features/mint-nft-drop/Steps/SelectAmount/SelectAmount.module.scss
