+++ src/features/nft-drop-raffle/WaitlistRegistration/validation.ts
